requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
onnx>=1.15.0
onnxruntime>=1.16.0
onnxconverter-common>=1.14.0
torch-pruning>=1.3.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0
tensorboard>=2.14.0
tqdm>=4.66.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
huggingface_hub>=1.0.0
